Imeplement users edge function auth (postgresql INSERT still doesn't fire)
