+++ react_native_frontend/app/register.tsx
@@ -1,10 +1,20 @@
 import React, { useState } from "react";
-import { View, TextInput, TouchableOpacity, Text, Alert } from "react-native";
+import {
+  View,
+  TextInput,
+  TouchableOpacity,
+  Text,
+  Alert,
+  StyleSheet,
+} from "react-native";
 import { useRouter } from "expo-router";
-import { supabase } from "../src/config/supabaseClient";
+// import { supabase } from "../src/config/supabaseClient";
+import { SUPABASE_URL } from "@env";
+import { SUPABASE_ANON_KEY } from "@env";
+
 import { useTheme } from "../src/context/ThemeContext";
-import ScreenContainer from "../components/ScreenContainer";
-import styles from "../components/ScreenStyles";
+// const edgeFunctionUrl = `${SUPABASE_URL}/auth/v1/signup`; // Edge function URL
+const edgeFunctionUrl = `${SUPABASE_URL}/functions/v1/users/`; // Edge function URL
 
 export default function RegisterScreen() {
   const [email, setEmail] = useState("");
@@ -27,48 +37,51 @@ export default function RegisterScreen() {
       return;
     }
 
-    // Attempt to sign up the user
-    const { data, error } = await supabase.auth.signUp({
-      email,
-      password,
-    });
+    // Debugging: Log registration request data
+    console.log("Registering user:", { email, firstName, lastName, username });
 
-    const user = data?.user; // Accessing user from the data object
+    try {
+      // Attempt to sign up the user using Edge Function instead of embedded Supabase code
+      const response = await fetch(edgeFunctionUrl, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          "apikey": SUPABASE_ANON_KEY,
+        },
+        body: JSON.stringify({
+          email,
+          password,
+          name: `${firstName} ${lastName}`,
+          username,
+        }),
+      });
 
-    // Debugging output
-    console.log("Sign-up response:", { user, error });
-    // // Note: delete this after figuring out why the uuid isn't being inserted in the users entry
-    // console.log("User object:", user, error);
+      // Debugging: Log response status and data
+      console.log("Response status:", response.status);
+      const data = await response.json();
+      console.log("Response data:", data);
 
-    if (error) {
-      console.error("Sign-up error:", error);
-      Alert.alert("Registration Error", error.message);
+      if (!response.ok) {
+        Alert.alert("Registration Error", data.error || "Failed to register");
+      } else {
+        Alert.alert("Success", "Account created! Please log in.");
+        router.push("/login");
+      }
+    } catch (error) {
+      console.error("Error during registration:", error);
+      Alert.alert("Error", "Something went wrong. Please try again.");
+    } finally {
       setLoading(false);
-      return; // todo: double check if returning here is the best approach
-    }
-
-    // Insert new user data (with uuid) into PostgreSQL database
-    // todo: use deployed Supabase Edge Functions instead of embedded Supabase code
-    const { error: dbError } = await supabase.from("users").insert([
-      {
-        name: `${firstName} ${lastName}`,
-        username, // Include username
-        uuid: user?.id, // Using the Supabase Auth UUID
-      },
-    ]);
-
-    if (dbError) {
-      Alert.alert("Database Error", dbError.message);
-      setLoading(false);
-    } else {
-      Alert.alert("Success", "Account created! Please log in.");
-      setLoading(false);
-      router.push("/login");
     }
   };
 
   return (
-    <ScreenContainer>
+    <View
+      style={[
+        styles.container,
+        theme === "dark" ? styles.darkContainer : styles.lightContainer,
+      ]}
+    >
       <Text
         style={[
           styles.appName,
@@ -82,7 +95,6 @@ export default function RegisterScreen() {
         style={[
           styles.formContainer,
           theme === "dark" ? styles.darkForm : styles.lightForm,
-          { alignItems: "center", paddingVertical: 20 },
         ]}
       >
         <Text
@@ -116,7 +128,7 @@ export default function RegisterScreen() {
         <TextInput
           placeholder="Username"
           placeholderTextColor={theme === "dark" ? "#999" : "#999"}
-          value={username}
+          value={username} // New username input
           onChangeText={setUsername}
           style={[
             styles.input,
@@ -161,7 +173,6 @@ export default function RegisterScreen() {
           style={[
             styles.button,
             theme === "dark" ? styles.darkButton : styles.lightButton,
-            { width: "45%" },
           ]}
           onPress={handleRegister}
           disabled={loading}
@@ -179,6 +190,97 @@ export default function RegisterScreen() {
           </Text>
         </TouchableOpacity>
       </View>
-    </ScreenContainer>
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    paddingHorizontal: 20,
+  },
+  darkContainer: {
+    backgroundColor: "#121212",
+  },
+  lightContainer: {
+    backgroundColor: "#f7f9fc",
+  },
+  appName: {
+    fontSize: 26,
+    fontWeight: "bold",
+    marginBottom: 20,
+  },
+  darkAppName: {
+    color: "#b05600",
+  },
+  lightAppName: {
+    color: "#f48c42",
+  },
+  darkText: {
+    color: "#fff",
+  },
+  lightText: {
+    color: "#000",
+  },
+  formContainer: {
+    width: "90%",
+    maxWidth: 400,
+    padding: 20,
+    borderRadius: 10,
+    shadowColor: "#000",
+    shadowOffset: { width: 0, height: 4 },
+    shadowOpacity: 0.2,
+    shadowRadius: 5,
+  },
+  darkForm: {
+    backgroundColor: "#333",
+  },
+  lightForm: {
+    backgroundColor: "#fff",
+  },
+  title: {
+    fontSize: 22,
+    fontWeight: "bold",
+    marginBottom: 20,
+    textAlign: "center",
+  },
+  input: {
+    borderWidth: 1,
+    padding: 10,
+    marginBottom: 15,
+    borderRadius: 8,
+    fontSize: 16,
+  },
+  darkInput: {
+    borderColor: "#666",
+    color: "#fff",
+  },
+  lightInput: {
+    borderColor: "#ccc",
+    color: "#000",
+  },
+  button: {
+    backgroundColor: "#f48c42",
+    paddingVertical: 12,
+    borderRadius: 30,
+    alignItems: "center",
+    marginBottom: 20,
+  },
+  darkButton: {
+    backgroundColor: "#b05600",
+  },
+  lightButton: {
+    backgroundColor: "#f48c42",
+  },
+  buttonText: {
+    color: "#fff",
+    fontSize: 16,
+    fontWeight: "600",
+  },
+  linkText: {
+    fontSize: 14,
+    textAlign: "center",
+  },
+});
