+++ supabase/.env.edge
@@ -1,0 +1,3 @@
+SUPABASE_SERVICE_ROLE_KEY="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InRqZnBuYnp3ZHprY3Zhc3lxYnlrIiwicm9sZSI6InNlcnZpY2Vfcm9sZSIsImlhdCI6MTcyOTMzNDU3OCwiZXhwIjoyMDQ0OTEwNTc4fQ.7QNQIs2vT0IMz0VxA7Cqu4k0sZdcsW9TCZOoSJePcvk"
+SUPABASE_URL=https://tjfpnbzwdzkcvasyqbyk.supabase.co
+SUPABASE_ANON_KEY="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InRqZnBuYnp3ZHprY3Zhc3lxYnlrIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MjkzMzQ1NzgsImV4cCI6MjA0NDkxMDU3OH0.tqG-gbdBcE2qhg0JeV5H1Aa16qfI5Ebs1-aSqggu3Gc"
