+++ supabase/functions/users/index.ts
@@ -2,19 +2,20 @@ import { config } from "https://deno.land/x/dotenv/mod.ts";
 
 // Load env variables
 // Note: I used eslint-disable-next-line no-unused-vars
-const env = config({ path: "../../.env.supabase" });
+// const env = config({ path: "../../.env.supabase" });
 
-console.log("All environment variables:", Deno.env.toObject());
-
-// Set up environment variables
 const supabaseUrl = Deno.env.get("SUPABASE_URL");
 const supabaseAnonKey = Deno.env.get("SUPABASE_ANON_KEY");
-// Todo: add supabaseServiceKey equivalent to localbackend users route(s) too
+// Todo: add supabaseServiceKey equivalent to localbackend users routes too
+// Service key for authenticated requests.
+// const supabaseServiceKey = Deno.env.get("SUPABASE_SERVICE_KEY");
 const supabaseServiceKey = Deno.env.get("SUPABASE_SERVICE_ROLE_KEY");
 
+
+
 // For testing
-console.log("SUPABASE_ANON_KEY:", supabaseAnonKey);
 console.log("SUPABASE_URL:", supabaseUrl);
+console.log("SUPABASE_ANON_KEY:", supabaseAnonKey);
 console.log("SUPABASE_SERVICE_KEY:", supabaseServiceKey);
 
 // Validate env variables
@@ -60,6 +61,24 @@ const validateJWT = async (token: string) => {
 };
 
 const handleRequest = async (req: Request) => {
+  // CORS headers to be added to all responses
+  // Todo: That means all responses for all routes
+  const corsHeaders = {
+    "Content-Type": "application/json",
+    "Access-Control-Allow-Origin": "*", // Replace '*' with the frontend domain once its deployed (for security reasons)
+    "Access-Control-Allow-Methods": "GET, POST, PATCH, DELETE, OPTIONS", // Allow the HTTP methods
+    "Access-Control-Allow-Headers": "Content-Type, Authorization, apikey", // Allow headers for the request
+  };
+
+  // Handle preflight OPTIONS requests
+  if (req.method === "OPTIONS") {
+    console.log("Handling OPTIONS request");
+    return new Response(null, {
+      status: 204, // No Content for OPTIONS request
+      headers: corsHeaders,
+    });
+  }
+
   const url = new URL(req.url);
   const path = url.pathname.split("/");
   const id = path.pop();
@@ -82,16 +101,19 @@ const handleRequest = async (req: Request) => {
         return await updateUser(id, await req.json());
 
       case "DELETE":
-        return await deleteUser(id);
+        return await deleteUser(id); // todo: the code works, but this probably still needs to be fixed
 
       default:
-        return new Response("Method Not Allowed", { status: 405 });
+        return new Response("Method Not Allowed", {
+          status: 405,
+          headers: corsHeaders,
+        });
     }
   } catch (error) {
     console.error("Internal Error:", error);
     return new Response(JSON.stringify({ error: error.message }), {
       status: 500,
-      headers: { "Content-Type": "application/json" },
+      headers: corsHeaders,
     });
   }
 };
@@ -104,7 +126,7 @@ const getUsers = async () => {
   });
   const data = await handleResponse(response);
   return new Response(JSON.stringify(data), {
-    headers: { "Content-Type": "application/json" },
+    headers: corsHeaders,
   });
 };
 
@@ -116,7 +138,7 @@ const getUser = async (id: string) => {
   );
   const data = await handleResponse(response);
   return new Response(JSON.stringify(data), {
-    headers: { "Content-Type": "application/json" },
+    headers: corsHeaders,
   });
 };
 
@@ -132,7 +154,7 @@ const createUser = async (body: {
       JSON.stringify({
         error: "You must enter a name, username, email, and password",
       }),
-      { status: 400 }
+      { status: 400, headers: corsHeaders }
     );
   }
 
@@ -154,8 +176,9 @@ const createUser = async (body: {
     );
   }
 
-  // Insert user data into PostgreSQL users table
+  // After successful signup, insert user details into the database
   const { user } = authData;
+  // Test if there is no user ID to be found
   if (!user || !user.id) {
     return new Response(
       JSON.stringify({ error: "User creation failed, no user ID" }),
@@ -173,7 +196,8 @@ const createUser = async (body: {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
-      "Authorization": `Bearer ${supabaseServiceKey}`, // Service Role Key for database insert
+      "Authorization": `Bearer ${supabaseServiceKey}`, // Use Supabase-granted access token
+      // "Authorization": `Bearer ${authData.access_token}`, // Use Supabase-granted access token
     },
     body: JSON.stringify({
       name: body.name,
@@ -185,6 +209,7 @@ const createUser = async (body: {
 
   const data = await handleResponse(dbResponse);
 
+  // Check for errors from the postgreSQL database
   if (!dbResponse.ok) {
     return new Response(
       JSON.stringify({
@@ -196,7 +221,7 @@ const createUser = async (body: {
 
   return new Response(JSON.stringify(data), {
     status: 201,
-    headers: { "Content-Type": "application/json" },
+    headers: corsHeaders,
   });
 };
 
@@ -205,7 +230,7 @@ const updateUser = async (id: string | undefined, body: { name: string }) => {
   if (!body.name) {
     return new Response(
       JSON.stringify({ error: "You must provide a name to update" }),
-      { status: 400 }
+      { status: 400, headers: corsHeaders }
     );
   }
 
@@ -227,13 +252,13 @@ const updateUser = async (id: string | undefined, body: { name: string }) => {
     const jsonData = JSON.parse(dataText);
     return new Response(JSON.stringify(jsonData), {
       status: 200,
-      headers: { "Content-Type": "application/json" },
+      headers: corsHeaders,
     });
   } else {
     // Case: response is empty
     return new Response(
       JSON.stringify({ message: "User updated successfully." }),
-      { status: 200 }
+      { status: 200, headers: corsHeaders }
     );
   }
 };
@@ -250,7 +275,7 @@ const deleteUser = async (id: string) => {
     throw new Error(errorData);
   }
 
-  return new Response(null, { status: 204 });
+  return new Response(null, { status: 204, headers: corsHeaders });
 };
 
 // Start the server
